+++ src/AdventOfCode2019/IO/IBufferedInputProvider.cs
@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Numerics;
 using System.Text;
 
 namespace AdventOfCode2019.IO
 {
     public interface IBufferedInputProvider : IInputProvider
     {
-        void AddInputValue(int value);
+        void AddInputValue(BigInteger value);
     }
 }
