+++ src/AdventOfCode2019/IO/IInputProvider.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Numerics;
 using System.Text;
 
 namespace AdventOfCode2019.IO
@@ -7,6 +8,6 @@ namespace AdventOfCode2019.IO
     public interface IInputProvider
     {
         bool HasInput();
-        int GetInput();
+        BigInteger GetInput();
     }
 }
