+++ src/AdventOfCode2019/IO/IOutputListener.cs
@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Numerics;
 using System.Text;
 
 namespace AdventOfCode2019.IO
 {
     public interface IOutputListener
     {
-        void SendOutput(int value);
+        void SendOutput(BigInteger value);
     }
 }
