+++ src/AdventOfCode2019/IO/ListOutputListener.cs
@@ -1,14 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Numerics;
 using System.Text;
 
 namespace AdventOfCode2019.IO
 {
     public class ListOutputListener : IOutputListener
     {
-        public List<int> Values = new List<int>();
+        public List<BigInteger> Values = new List<BigInteger>();
 
-        public void SendOutput(int value)
+        public void SendOutput(BigInteger value)
         {
             Values.Add(value);
         }
