+++ src/AdventOfCode2019Test/Challenges/Day2Test.cs
@@ -1,6 +1,7 @@
 ﻿using AdventOfCode2019.Challenges.Day2;
 using System;
 using System.Collections.Generic;
+using System.Numerics;
 using System.Text;
 using Xunit;
 
@@ -11,8 +12,8 @@ namespace AdventOfCode2019Test.Challenges
         [Fact]
         public void GetDay2Part1AnswerTest()
         {
-            int expected = 11590668;
-            int actual = Day2.GetDay2Part1Answer();
+            BigInteger expected = 11590668;
+            BigInteger actual = Day2.GetDay2Part1Answer();
             Assert.Equal(expected, actual);
         }
